+++ packages/openapi-parser/src/utils/index.ts
@@ -6,7 +6,7 @@ export * from './getListOfReferences'
 export * from './getSegmentsFromPath'
 export * from './isFilesystem'
 export * from './isObject'
-export * from './load'
+export * from './load/load'
 export * from './loadFiles'
 export * from './normalize'
 export * from './resolve'
