+++ packages/openapi-parser/src/utils/isJson.ts
@@ -1,0 +1,9 @@
+export function isJson(value: string) {
+  try {
+    JSON.parse(value)
+
+    return true
+  } catch {
+    return false
+  }
+}
